- Add liked & disliked tables - Create Liked & Disliked Classes - Create addLikedShops, addDislikedShops, deleteLikedShops & deleteDislikedShops methods. - Add routes

+++ routes/web.php
@@ -15,14 +15,29 @@ $router->get('/', function () use ($router) {
     return $router->app->version();
 });
 
+
 $router->group(['prefix' => 'api'], function () use ($router) {
-  $router->get('users',  ['uses' => 'UserController@showAllUsers']);
 
-  $router->get('users/{id}', ['uses' => 'UserController@showOneUser']);
+	$router->post('login',  ['uses' => 'UserController@login']);
 
-  $router->post('users', ['uses' => 'UserController@create']);
+	$router->get('users',  [ 'uses' => 'UserController@showAllUsers']);
 
-  $router->delete('users/{id}', ['uses' => 'UserController@delete']);
+	$router->get('users/{id}', ['uses' => 'UserController@showOneUser']);
 
-  $router->put('users/{id}', ['uses' => 'UserController@update']);
+	$router->post('users', ['uses' => 'UserController@create']);
+
+	$router->delete('users/{id}', ['middleware' => 'auth', 'uses' => 'UserController@delete']);
+
+	$router->put('users/{id}', ['middleware' => 'auth', 'uses' => 'UserController@update']);
+
+	$router->get('users/liked/{id}', ['middleware' => 'auth', 'uses' => 'UserController@showUserLikedShops']);
+	$router->get('users/disliked/{id}', ['middleware' => 'auth', 'uses' => 'UserController@showUserLislikedShops']);
+
+	$router->post('users/liked/{id}/{shop_id}', ['middleware' => 'auth', 'uses' => 'UserController@addLiked']);
+	$router->post('users/disliked/{id}/{shop_id}', ['middleware' => 'auth', 'uses' => 'UserController@addDisliked']);
+
+	$router->delete('users/liked/{id}', ['middleware' => 'auth', 'uses' => 'UserController@deleteLiked']);
+	$router->delete('users/disliked/{id}', ['middleware' => 'auth', 'uses' => 'UserController@deleteDisliked']);
+
+	
 });
